+++ src/Task/Prep.php
@@ -1,0 +1,91 @@
+<?php
+
+/**
+ * @package Effigy
+ * @license http://opensource.org/licenses/MIT
+ */
+
+declare(strict_types=1);
+
+namespace DecodeLabs\Effigy\Task;
+
+use DecodeLabs\Clip\Task;
+use DecodeLabs\Effigy;
+use DecodeLabs\Terminus as Cli;
+
+class Prep implements Task
+{
+    public function execute(): bool
+    {
+        // Update
+        if (!Effigy::run('update')) {
+            return false;
+        }
+
+
+        Cli::newLine();
+        Cli::newLine();
+
+
+        // Analysis
+        Cli::info('Clear PHPStan cache');
+
+        if (!Effigy::run('analyze', '--clear')) {
+            return false;
+        }
+
+        Cli::newLine();
+        Cli::newLine();
+
+        Cli::info('Run PHPStan analysis');
+
+        if (!Effigy::run('analyze')) {
+            return false;
+        }
+
+
+        Cli::newLine();
+        Cli::newLine();
+
+        // Standards
+        Cli::info('Ensure proper code formatting');
+
+        if (!Effigy::run('format')) {
+            return false;
+        }
+
+        Cli::newLine();
+        Cli::newLine();
+
+        // Lint
+        Cli::info('Ensure all files are syntax compliant');
+
+        if (!Effigy::run('lint')) {
+            return false;
+        }
+
+
+        Cli::newLine();
+        Cli::newLine();
+
+        // EC Lint
+        Cli::info('Check for editorconfig issues');
+
+        if (!Effigy::run('eclint')) {
+            return false;
+        }
+
+
+        Cli::newLine();
+        Cli::newLine();
+
+        // Non ascii
+        Cli::info('Check for non-ASCII characters');
+
+        if (!Effigy::run('check-non-ascii')) {
+            return false;
+        }
+
+        return true;
+    }
+}