+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Switched to Clip for process initialisation
+* Renamed Command to Task
+* Register Controller as Veneer facade
+* Passthrough to available vendor bins
 * Create src folder during init-package
 
 ## v0.1.14 (2022-11-22)